etl/requirements.txt
kafka-python==2.0.2
redis==4.5.5
requests==2.32.3
